[SPARK-46391][PS][TESTS] Reorganize `ExpandingParityTests`
### What changes were proposed in this pull request?
Reorganize `ExpandingParityTests`

### Why are the changes needed?
to make the test more consistent with pandas

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44332 from zhengruifeng/ps_test_expanding.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
