+++ dev/sparktestsupport/modules.py
@@ -727,7 +727,11 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.test_dataframe_conversion",
         "pyspark.pandas.tests.test_dataframe_spark_io",
         "pyspark.pandas.tests.test_default_index",
-        "pyspark.pandas.tests.test_expanding",
+        "pyspark.pandas.tests.window.test_expanding",
+        "pyspark.pandas.tests.window.test_expanding_adv",
+        "pyspark.pandas.tests.window.test_expanding_error",
+        "pyspark.pandas.tests.window.test_groupby_expanding",
+        "pyspark.pandas.tests.window.test_groupby_expanding_adv",
         "pyspark.pandas.tests.test_extension",
         "pyspark.pandas.tests.window.test_ewm_error",
         "pyspark.pandas.tests.window.test_ewm_mean",
@@ -1135,7 +1139,11 @@ pyspark_pandas_connect_part2 = Module(
         "pyspark.pandas.tests.connect.window.test_parity_groupby_rolling",
         "pyspark.pandas.tests.connect.window.test_parity_groupby_rolling_adv",
         "pyspark.pandas.tests.connect.window.test_parity_groupby_rolling_count",
-        "pyspark.pandas.tests.connect.test_parity_expanding",
+        "pyspark.pandas.tests.connect.window.test_parity_expanding",
+        "pyspark.pandas.tests.connect.window.test_parity_expanding_adv",
+        "pyspark.pandas.tests.connect.window.test_parity_expanding_error",
+        "pyspark.pandas.tests.connect.window.test_parity_groupby_expanding",
+        "pyspark.pandas.tests.connect.window.test_parity_groupby_expanding_adv",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_groupby_rolling",
         "pyspark.pandas.tests.connect.computation.test_parity_missing_data",
         "pyspark.pandas.tests.connect.groupby.test_parity_index",
