+++ python/pyspark/pandas/tests/connect/window/test_parity_expanding_adv.py
@@ -16,19 +16,21 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_expanding import ExpandingTestsMixin
+from pyspark.pandas.tests.window.test_expanding_adv import ExpandingAdvMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
-from pyspark.testing.pandasutils import PandasOnSparkTestUtils, TestUtils
+from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
-class ExpandingParityTests(
-    ExpandingTestsMixin, PandasOnSparkTestUtils, TestUtils, ReusedConnectTestCase
+class ExpandingAdvParityTests(
+    ExpandingAdvMixin,
+    PandasOnSparkTestUtils,
+    ReusedConnectTestCase,
 ):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_expanding import *  # noqa: F401
+    from pyspark.pandas.tests.connect.window.test_parity_expanding_adv import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
